Make npoles unsigned

+++ libformant/jkFormant.h
@@ -29,7 +29,7 @@ typedef struct {   /* structure to hold raw LPC analysis data */
     double f0;     /* fundamental frequency estimate for this frame */
     double pv;		/* probability that frame is voiced */
     double change; /* spec. distance between current and prev. frames */
-    short npoles; /* # of complex poles from roots of LPC polynomial */
+    size_t npoles; /* # of complex poles from roots of LPC polynomial */
     double *freq;  /* array of complex pole frequencies (Hz) */
     double *band;  /* array of complex pole bandwidths (Hz) */
 } pole_t;
